refactor(toolshed): don't import the app graph in the --background parent
The toolshed's `--background` mode has the foreground process spawn a
second copy of itself as the real server, wait for that copy to report it
has bound its port, and then exit. The foreground process is only a
launcher: it decides which half of the launch it is, spawns the child, and
returns once the child is listening.

The entry module imported the whole application graph statically, at the
very top:

    import app from "@/app.ts";

Importing `@/app.ts` runs startup side effects before the process has
decided whether it is the launcher or the server. It opens the memory-store
provider and fetches the gateway model list over the network. So the
`--background` launcher ran all of that work too, redundantly: the memory
store was briefly opened by two processes, and a failure while importing the
app killed the launcher before it could spawn the child or surface anything
through the child's log file. (Continuous integration masks this because it
sets an empty gateway URL, which suppresses the network fetch.)

This splits the entry module in two so the app graph is imported only on the
path that actually serves.

`index.ts` is now a thin launcher. It imports only `@/background.ts`, which
pulls in nothing heavy — just the standard-library path helpers and Deno
APIs. In its `import.meta.main` block it reads the background-log environment
variable and classifies the command line. When this process is the
background parent (the caller asked to background it and no log variable
marks it as the server half), it spawns the child and waits, and never
imports the app graph. Otherwise — a foreground launch, or the server half
of a background launch — it dynamically imports `@/server.ts` and starts the
server, applying the child-mode console redirect, uncaught-error logging, and
readiness marker when this process is the server half.

The new `server.ts` holds the server itself: the `import app from "@/app.ts"`,
the shared `runtime` binding and its initialization, the abort controller and
signal handlers, the graceful shutdown, and `startServer`. Importing this
module is what runs the startup side effects, so only the serving path
reaches it.

The `runtime` binding moved from `index.ts` to `server.ts`; the handler files
that read it now import it from `@/server.ts`. There is a cycle here — the
server imports the app, the app imports the handlers, and the handlers import
`runtime` back from the server — but it is the same shape the entry module
already had, and it works because `runtime` is read when a request is served,
not when the module is imported, so the live binding is populated by the time
any handler runs.

`index.ts` remains the entry module. The child command the parent builds
still points at it: the compiled binary re-runs itself, and a `deno run`
launch re-runs Deno against `index.ts`. The child boots through `index.ts`,
takes the serving branch, and dynamically imports `server.ts`. `deno compile`
bundles that dynamic import because its specifier is a string literal, so the
compiled binary carries `server.ts` and the background child starts from it.

Because the runtime binding and the runtime construction around it moved out
of `index.ts`, references that named `index.ts` as their home are updated to
`server.ts`: the wiring diagram in the experimental-options registry, the
server-hosts row of the CFC enforcement matrix, three test-file comments that
explain the import cycle, and the module listing in the toolshed README.
