+++ docs/development/EXPERIMENTAL_OPTIONS.md
@@ -779,7 +779,7 @@ Server Process (Deno)
   |
   +-- runner/runtime-presets.ts  --> experimentalOptionsFromEnv(Deno.env.get)
   +-- toolshed/runtime-options.ts --> runtimePresets.productionServer({ experimental, ... })
-  +-- toolshed/index.ts           --> new Runtime(toolshedRuntimeOptions(...))
+  +-- toolshed/server.ts          --> new Runtime(toolshedRuntimeOptions(...))
 ```
 
 The background piece service and the CLI use the same mapping and the same
