+++ packages/toolshed/routes/ingest/ingest.routes.test.ts
@@ -9,7 +9,7 @@ if (env.ENV !== "test") {
 
 // Smoke tests: confirm POST /api/ingest/:id is mounted and its transport-level
 // branches behave. Driven through the real `app` (not a freshly-mounted router)
-// because the handler imports the `runtime` singleton from `@/index.ts`, which
+// because the handler imports the `runtime` singleton from `@/server.ts`, which
 // is uninitialized under test — so any path that reaches storage yields 502,
 // which is itself the storage-error contract. The full auth + validation
 // contract is unit-tested against a real runtime in ingest.utils.test.ts
