+++ packages/toolshed/routes/integrations/oauth2-common/oauth2-common.handlers.ts
@@ -2,7 +2,7 @@ import type { Tokens } from "@cmd-johnson/oauth2-client";
 import type { Context } from "@hono/hono";
 import { getLogger } from "@commonfabric/utils/logger";
 import { setBGPiece } from "@commonfabric/background-piece";
-import { runtime } from "@/index.ts";
+import { runtime } from "@/server.ts";
 import { OAuth2TokenSchema } from "@commonfabric/runner";
 import type { JSONSchema } from "@commonfabric/runner";
 import type {
