+++ packages/toolshed/routes/webhooks/webhooks.utils.ts
@@ -1,6 +1,6 @@
 import { getLogger } from "@commonfabric/utils/logger";
 import { sha256 } from "@/lib/sha2.ts";
-import { runtime } from "@/index.ts";
+import { runtime } from "@/server.ts";
 import { identity } from "@/lib/identity.ts";
 import { WebhookConfigSchema } from "@commonfabric/runner";
 import {
